services/f5/requirements.txt
fastapi
uvicorn
soundfile
numpy
omegaconf
hydra-core
ruaccent
f5-tts
jieba
pypinyin